// src/shared/enums/messages/note.ts
export enum MessagesNote {
    UPDATE = "Nota modificada con éxito",
    DELETE = "Nota eliminada",
    DELETES = "Notas eliminadas",
    CREATED = "Nota creada con éxito",
    BAD_IMAGE = "Solo se permiten archivos de imagen",
    NOT_FOUND = "Nota no encontrada",
    FILE_ERROR = "Error al analizar el archivo",
    DUPLICATED = "Error la nota que intentas crear ya existe",
    FILE_MALWARE = "Archivo con virus",
    DELETE_FAILED = "Error de eliminación",
    FILE_NOT_FOUND = "Archivo no encontrado"
}